+++ src/lib.rs
@@ -1,6 +1,7 @@
 mod conversions;
 mod egraph;
 mod error;
+mod extract;
 mod py_object_sort;
 mod serialize;
 mod termdag;
@@ -26,10 +27,13 @@ fn bindings(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<crate::py_object_sort::PyObjectSort>()?;
     m.add_class::<crate::serialize::SerializedEGraph>()?;
     m.add_class::<crate::egraph::EGraph>()?;
+    m.add_class::<crate::egraph::Value>()?;
     m.add_class::<crate::error::EggSmolError>()?;
     m.add_class::<crate::termdag::TermDag>()?;
     m.add_class::<crate::conversions::UserDefinedCommandOutput>()?;
     m.add_class::<crate::conversions::Function>()?;
+    m.add_class::<crate::extract::Extractor>()?;
+    m.add_class::<crate::extract::CostModel>()?;
     crate::conversions::add_structs_to_module(m)?;
     crate::conversions::add_enums_to_module(m)?;